requirements.txt
babel==2.16.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1